First commit: cleaning unncessary files, adding typescript features, search.tsx component created
