+++ src/App.tsx
@@ -1,0 +1,11 @@
+import Search from "./components/Search";
+
+function App() {
+  return (
+    <div>
+      <Search />
+    </div>
+  );
+}
+
+export default App;
