+++ src/components/Search.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Search = () => {
+  return <div>input to come</div>;
+};
+
+export default Search;
